MM-8604: handle SERVER_VERSION_CHANGED and CONFIG_CHANGED events
This pairs up with the corresponding mattermost-redux changes to split
apart these events. For now, we still assume a configuration change when
detecting a server version change, given that the app may run with older
mattermost servers. In the future, we could fine-tine or remove this duplicate
config/license loading.

TODO: Depend on mattermost-redux changes in yarn.lock.