add back "a modified version of..." notices for dependencies that are modified by Mattermost
This is determined by looking at the requested version string in package.json. If the requested version is of the form "github:.../...(#...)", where the GitHub repo is something other than the official upstream repo (as given in the NPM registry entry), then we flag the dependency as being "a modified version".