Update NOTICE.txt (#2163)
* Update NOTICE.txt

- Add new dependencies (jsc-android, moment-timezone, react-native-calendars, react-native-drawer-layout, react-navigation, rn-placeholder)
- Remove no-longer-used dependencies (react-native-drawer, react-native-fast-image)
- Update dependency with new name (TinyColor->tinycolor2)
- Update homepage and project owner info for all dependencies
- Add SPDX open-source license IDs
- Alphabetize dependency list so future update diffs will be smaller

* add back "a modified version of..." notices for dependencies that are modified by Mattermost

This is determined by looking at the requested version string in package.json. If the requested version is of the form "github:.../...(#...)", where the GitHub repo is something other than the official upstream repo (as given in the NPM registry entry), then we flag the dependency as being "a modified version".