MM-19599 Updated Channel Nav Bar height
During the landscape changes from the previous release, I had added 8px of padding to the iOS landscape, which caused this issue to occur. This change reverts the added padding so that the height matches across the screens for the navigation bar.

To note, react-native-navigation is used for the info/thread headers and there is no way to adjust the height of this without replacing the header entirley.