MM-23135 Remove double-handling of toggle switch state
Whether the Mute, Favorite, Ignore toggles are set on or off is already being governed by the parent controlling component. Whenever any of their respective states change, the parent updates the child's props, and the switches re-render with the correct switch settings in props anyway.

`getDerivedStateFromProps` is called on component creation and then whenever the parent component's state is updated and a re-render of all children is triggered, **regardless of whether the props are different from before **

The `this.setState` in each of the toggle switch handlers was already taking care of keeping component state healthy. With `getDerivedStateFromProps`, state was being "double-managed" resulting in toggle switches sometimes jumping between on and off, well after it was touched.