MM-31322: handle _redirect permalinks
In the webapp, [there is support](https://github.com/mattermost/mattermost-webapp/blob/7c5a2e3bfdf7bb66ac88642e3b426152bbc1fceb/components/root/root.jsx#L345-L352) for a special `_redirect` team name in certain contexts that simply means, "the current team". It's used in the context of redirecting to the integrations backstage (not relevant for mobile), for also for resolving post permalinks.

This simplifies certain kinds of integrations which no longer have to know a team name to render a permanent link to a post. An example link looks like:

    http://localhost:8065/_redirect/pl/e3sxrxtwh78jmxawsaqfemxoew

This PR adds support for same in mobile, unintentionally missed on the first round of implementation.

Fixes: https://mattermost.atlassian.net/browse/MM-31322