MM-35017: participants ordering and inclusion
Participants no longer include the thread starter,
thread starter is included only if the user has replied.

This commit changes the ordering of participants to reverse so that the
latest replier is placed first and so forth.

Also the thread starter is not included in the channel view, but is
included in the global threads view.