Remove reliable websockets configuration
The server side configuration will be removed
and the feature will always be enabled now.

Hence removing the if conditions and updating
the code to assume that reliable websockets
is always true.

For compatibility with older clients, the server
will always return true for this config setting
unless we reach to an unsupported version.

https://community-daily.mattermost.com/boards/workspace/zyoahc9uapdn3xdptac6jb69ic/285b80a3-257d-41f6-8cf4-ed80ca9d92e5/495cdb4d-c13a-4992-8eb9-80cfee2819a4/7m95frqbk9o8zxin6xo9c7iusoh

```release-note
NONE
```